refactor: Moving Telemetry to Observability
