+++ internal/observability/telemetry/config.go
@@ -1,4 +1,4 @@
-package reconciler
+package telemetry
 
 import (
 	"fmt"
