+++ internal/observability/telemetry/configmap.go
@@ -1,4 +1,4 @@
-package reconciler
+package telemetry
 
 import (
 	"context"
