+++ internal/observability/telemetry/telemetry_chart_test.go
@@ -1,4 +1,4 @@
-package reconciler
+package telemetry
 
 import (
 	"os/exec"
